Use .pth and ld.so.conf files instead of env vars
Add the GR Python packages and the libraries installed on the GR prefix
using a Python .pth file and a conf file on ld.so.conf instead of
defining the PYTHONPATH and LD_LIBRARY_PATH env vars. The latter, in
particular, when defined, breaks some tests executed by "make test", as
it disables other dynamic linker paths required to run them. By adding
the conf file to ld.so.conf.d/, the GR path is simply added to the list
of paths, as desired.